Fix more linting and tests in debugger
